+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2020-10-12
+- Added real-time notifications new listeners and agents [#49](https://github.com/BC-SECURITY/Starkiller/pull/49) 
+- Log out the user when a connection can't be established to the Empire server
+
 ## [1.3.2] - 2020-07-29
 - "Listener" option on modules now shows up as a dropdown instead of a textfield
 - Correct some of the "cannot read property data of undefined" errors